final pass edits
notes:

General:

Noticed looking at them in Jupyter that after a ``` python code block within a “hint” section, the blank line doesn’t display - so have added an extra <br> in those cases so that it displays as intended within Jupyter.

0

“But I wanted to mention it here so you're not confused when you come across Jupyter Notebooks where you can't edit cells.”

Removed “I” to be consistent with the rest of the course (also later instance in lesson 1)

—

“Try out what you've learned before in the code cell below. Select the cell below, enter Edit Mode with Enter, “

Changed this (and added a note above) since you don’t need to hit enter to edit a code cell.

—
 Added a ref to the kernel restart thing

1

Added the section about logging in and out again / kernel restart to the “Jupyter what?” Section - that was in ) and if that’s now not part of the course then we need it here

—

Changed mention of “reserved words” since we actually introduce that term in lesson 2 now

3

I moved the added bit about operators and parentheses to lesson 4. I thought the example was confusing at this point cause we hadn’t covered conditionals yet, so I re-wrote it in lesson 4 with a code block so they can actually see the difference. (It also broke up the flow here)

4

Removed LLM mention because we don’t want to encourage group B to use them.

6

Removed mention of “AI assistants”, for the same reason as above

8

“After the variable name of the iterating item comes the keyword `in`, followed by the element to be iterated upon. In this case, we're iterating on a list. We did this by using the variable name in which the list is stored. But you could also iterate on a list like this:”

Mark suggested this was overly complicated so changed to “In the above example, we iterated on the list by using the variable name in which the list is stored. But you could also iterate on a list like this:”